changed telegram_id from Integer to BigInteger

+++ users/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -7,7 +7,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = "users"  # Define the table name
     id = Column(Integer, primary_key=True)  # Primary key
-    telegram_id = Column(Integer, unique=True)  # Unique Telegram ID
+    telegram_id = Column(BigInteger, unique=True)  # Unique Telegram ID
     username = Column(String)  # Telegram username
     has_access = Column(Boolean, default=False)  # Access status (default: False)
     subscription_end = Column(DateTime)  # Subscription end date
